requirements.txt
click
openai
python-chess
python-dotenv